requirements.txt
notion-client>=2.0.0
python-dotenv>=1.0.0
click>=8.0.0
